feat: 댓글 메타 조회 책임 분리
